+++ packages/audits/src/Api/selectors.ts
@@ -1,16 +1,16 @@
 import { createSelector } from 'reselect';
-import { AuditsState } from './interfaces';
+import { OldAuditsState as AuditsState } from '@frontegg/redux-store/audits/backward-compatibility';
 
 const filters = (state: AuditsState) => state.filters;
 const headerProps = (state: AuditsState) => state.headerProps;
 const predefinedFilters = (state: AuditsState) => state.predefinedFilters;
 
 export const filtersWithoutPredefined = createSelector(predefinedFilters, filters, (predefinedFilters, filters) =>
-  filters.filter((f) => !Object.keys(predefinedFilters).includes(f.key))
+  filters.filter((f: any) => !Object.keys(predefinedFilters).includes(f.key))
 );
 
 export const headerPropsWithoutPredefinedFilters = createSelector(
   headerProps,
   predefinedFilters,
-  (headerProps, predefinedFilters) => headerProps.filter((i) => !Object.keys(predefinedFilters).includes(i.name))
+  (headerProps, predefinedFilters) => headerProps.filter((i: any) => !Object.keys(predefinedFilters).includes(i.name))
 );