+++ packages/audits/src/components/AuditsTable.tsx
@@ -2,6 +2,7 @@ import React, { FC, useCallback, useEffect, useMemo } from 'react';
 import { TableFilter, TableSort } from '@frontegg/react-core';
 import { useAuditsState, useAuditsActions } from '../helpers/hooks';
 import { AuditsRawTable } from './AuditsRawTable';
+import { ContextHolder } from '@frontegg/rest-api';
 
 export const AuditsTable: FC = () => {
   const { isLoading, headerProps, rowsData, filters, total, virtualScroll } = useAuditsState(
@@ -14,13 +15,17 @@ export const AuditsTable: FC = () => {
       virtualScroll,
     })
   );
-  const { onPageChange, setDataSorting, setFilterData, initData } = useAuditsActions();
+  const { onPageChange, setDataSorting, setFilterData, initData, setVirtualScroll } = useAuditsActions();
 
   useEffect(() => {
+    const context = ContextHolder.getContext();
+    setVirtualScroll(!!context?.auditsOptions?.virtualScroll);
     initData();
   }, [initData]);
 
-  const dataFilters = useMemo(() => (filters ? filters.map((f) => ({ id: f.key, value: f.value })) : []), [filters]);
+  const dataFilters = useMemo(() => (filters ? filters.map((f: any) => ({ id: f.key, value: f.value })) : []), [
+    filters,
+  ]);
   const handlerPageChange = useCallback(
     (pageSize: number, pageIndex: number) => {
       onPageChange(pageIndex + 1);