+++ packages/auth/src/AccountDropdown/AccountPopupSection.tsx
@@ -14,7 +14,7 @@ export type AccountPopupSectionProps = {
 export const AccountPopupSection: FC<AccountPopupSectionProps> = (props) => {
   const { title, items } = props;
   return (
-    <div className='fe-account-popup-section' data-test-id={props.dataTestId}>
+    <div className='fe-account-popup-section' data-test-id={props.title}>
       {title && <div className='fe-account-popup-section__title'>{title}</div>}
       {items.map((item, index) => (
         <MenuItem key={index} withIcons={true} icon={item.icon} onClick={item.onClick} text={<>{item.title}</>} />