+++ packages/core/src/elements/Table/FeTableTBody.tsx
@@ -48,8 +48,40 @@ export const FeTableTBodyRow: FC<FeTableTBodyRowProps<any>> = (props: FeTableTBo
 };
 
 export const FeTableTBody: FC<FeTableTBodyProps<any>> = <T extends object>(props: FeTableTBodyProps<T>) => {
-  const { getTableBodyProps, prepareRow, rows, renderExpandedComponent, loading, pagination, onInfiniteScroll } = props;
+  const {
+    getTableBodyProps,
+    prepareRow,
+    rows,
+    renderExpandedComponent,
+    loading,
+    pagination,
+    onInfiniteScroll,
+    pageSize,
+  } = props;
   const { t } = useT();
+
+  const isInfiniteScroll = pagination === 'infinite-scroll';
+  const isFirstWaypoint = useMemo(() => rows.length <= (pageSize ?? 20), [pageSize, rows.length]);
+
+  const renderWaypoint = (index: number) => {
+    const itemsAfterWaypoint = 15;
+    const itemsAfterWaypointOnFirstRender = 4;
+    const waypoint = (
+      <Waypoint
+        onEnter={({ previousPosition }) => {
+          if (!loading && previousPosition !== 'above') {
+            onInfiniteScroll?.();
+          }
+        }}
+      />
+    );
+    if (isFirstWaypoint && index === rows.length - itemsAfterWaypointOnFirstRender) {
+      return waypoint;
+    }
+    if (!isFirstWaypoint && index === rows.length - itemsAfterWaypoint) {
+      return waypoint;
+    }
+  };
 
   return (
     <div
@@ -61,15 +93,7 @@ export const FeTableTBody: FC<FeTableTBodyProps<any>> = <T extends object>(props
       {rows.map((row, index) => (
         <React.Fragment key={row.id}>
           <FeTableTBodyRow prepareRow={prepareRow} row={row} renderExpandedComponent={renderExpandedComponent} />
-          {pagination === 'infinite-scroll' && index === Math.ceil(rows.length * 0.7) && (
-            <Waypoint
-              onEnter={({ previousPosition }) => {
-                if (!loading && previousPosition !== 'above') {
-                  onInfiniteScroll?.();
-                }
-              }}
-            />
-          )}
+          {isInfiniteScroll && renderWaypoint(index)}
         </React.Fragment>
       ))}