+++ packages/demo-saas/src/withFrontegg.tsx
@@ -1,32 +1,45 @@
-import React, { ComponentType } from 'react';
+import React, { ComponentType, useContext, useEffect, useRef } from 'react';
 import { ContextOptions, FronteggProvider, PluginConfig } from '@frontegg/react-core';
 import { AuthPlugin } from '@frontegg/react-auth';
 import { ConnectivityPlugin } from '@frontegg/react-connectivity';
 import { AuditsPlugin } from '@frontegg/react-audits';
 import { uiLibrary } from '@frontegg/react-elements-material-ui';
-const developmentHosts = ['localhost', 'local.frontegg.com'];
-const host =
-  developmentHosts.indexOf(window.location.hostname) !== -1
-    ? `${window.location.hostname}:8080`
-    : window.location.hostname;
+import { FronteggStoreContext } from '@frontegg/react-hooks';
+import { initialize } from '@frontegg/admin-portal';
 
 const contextOptions: ContextOptions = {
-  baseUrl: `https://maxx.stg.frontegg.com`,
+  baseUrl: `https://david.frontegg.com`,
   requestCredentials: 'include',
   auditsOptions: {
     virtualScroll: true,
   },
 };
 
-const plugins: PluginConfig[] = [
-  AuthPlugin(),
-  ConnectivityPlugin(),
-  // NotificationsPlugin(),
-  AuditsPlugin(),
-];
+const plugins: PluginConfig[] = [AuthPlugin(), ConnectivityPlugin(), AuditsPlugin()];
 
+const ConnectAdminPortal = ({ children }: any) => {
+  const appRef = useRef<any>(null);
+  const { store } = useContext(FronteggStoreContext);
+
+  useEffect(() => {
+    appRef.current =
+      store &&
+      initialize({
+        version: 'next',
+        contextOptions,
+        customLoader: true,
+        customLoginBox: true,
+        store,
+        cdn: 'http://localhost:5000',
+        usingFronteggReactCore: true,
+      } as any);
+  }, [store]);
+  return <>{children}</>;
+};
 export const withFrontegg = (Component: ComponentType<any>) => () => (
   <FronteggProvider debugMode context={contextOptions} plugins={plugins} uiLibrary={uiLibrary}>
-    <Component />
+    <ConnectAdminPortal>
+      <Component />
+    </ConnectAdminPortal>
   </FronteggProvider>
 );