+++ packages/react/src/index.ts
@@ -1,5 +1,7 @@
 export * from './FronteggProvider';
+export * from './AuthorizedContent';
 
 export { AdminPortal } from '@frontegg/admin-portal';
 export * from '@frontegg/react-hooks/auth';
 export * from '@frontegg/react-hooks/audits';
+export * from '@frontegg/react-hooks/connectivity';